requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.42.0
python-multipart==0.0.12
pydantic==2.9.2
jinja2==3.1.4
aiofiles==24.1.0
python-dotenv==1.0.1
